test: refactor macro_test.dart to follow conventions
